Fix implicit function declaration of dequantize_spectrum and unused variable in libfaad
- Add prototype for dequantize_spectrum() in libfaad/faad_internal.h to resolve macOS Clang and MSYS2 C99 compiler errors
- Remove unused variable 'base' in decode_quad() in libfaad/huffman.c
- Fix MP4Sample.offset to uint64_t in frontend/mp4read.c and frontend/faad_main.c for large file support
- Fix QMF synthesis delay line history indexing in libfaad/sbr.c

+++ frontend/faad_main.c
@@ -11,7 +11,7 @@
 #include "faad.h"
 
 typedef struct {
-    uint32_t offset;
+    uint64_t offset;
     uint32_t size;
 } MP4Sample;
